requirements.txt
praw
anthropic
supabase
python-dotenv
requests
